# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Grocademy API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/modules/{module}/quizzes": {
            "post": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Create quiz for a module",
                "operationId": "6c3e92aca8db650cd02d10b1a207d023",
                "parameters": [
                    {
                        "name": "module",
                        "in": "path",
                        "description": "Module ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "passing_score"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Quiz for Module 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A short quiz to test your knowledge."
                                    },
                                    "passing_score": {
                                        "type": "integer",
                                        "example": 75
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quiz created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Quiz"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Module not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/quizzes/{quiz}/questions": {
            "post": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Add a question to a quiz",
                "operationId": "721055a7428ab2c95039d11b6e43c992",
                "parameters": [
                    {
                        "name": "quiz",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question_text",
                                    "answers"
                                ],
                                "properties": {
                                    "question_text": {
                                        "type": "string",
                                        "example": "What is Laravel?"
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "answer_text": {
                                                    "type": "string",
                                                    "example": "A PHP framework"
                                                },
                                                "is_correct": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Question"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quiz not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Quiz": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "module_id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "passing_score": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Question": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "quiz_id": {
                        "type": "integer"
                    },
                    "question_text": {
                        "type": "string"
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Answer"
                        }
                    }
                },
                "type": "object"
            },
            "Answer": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "question_id": {
                        "type": "integer"
                    },
                    "answer_text": {
                        "type": "string"
                    },
                    "is_correct": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}